Moved playchecker
