+++ src/analysis/play-utilities/PlayChecker.cpp
@@ -13,36 +13,28 @@
 #include "functional"
 
 namespace rtt::ai::analysis {
+using namespace rtt::ai::world;
 
-bool PlayChecker::checkCurrentGameInvariants(rtt::ai::world::World *world, const rtt::ai::world::Field &field) { return true; }
+bool PlayChecker::checkCurrentGameInvariants(World *world, const Field *field) { return true; }
 
 /**
  * Determines what plays are viable given the current world, ref states and invariants/preconditions, and stores them in the validPlays vector
  * TODO: add lambda here, to make it faster and cleaner
  */
-void PlayChecker::determineNewPlays(rtt::ai::world::World *world, const rtt::ai::world::Field &field) {
+void PlayChecker::determineNewPlays(rtt::ai::world::World *world, const Field *field) {
     validPlays.clear();
-    for (auto const &play : allPlays) {
-        if (play->isValidPlay(world, field)) {
-            validPlays.push_back(play.get());
+    for (auto play : allPlays) {
+        if (play.isValidPlay(world, field)) {
+            validPlays.push_back(play);
         }
     }
 }
 
-bool PlayChecker::update(rtt::ai::world::World *world, const rtt::ai::world::Field &field) {
-    if (checkCurrentGameInvariants(world, field)) {
-        //        std::cout << "current play is still valid" << std::endl;
-        return true;
-    }
-    // Otherwise we select new plays for the playdecider by updating the validPlays vector of this object. This can then be accessed by the playdecider.
-    else {
-        std::cout << "current play is not valid anymore, calculating new valid plays" << std::endl;
+void PlayChecker::update(World *world, const Field *field) {
+    if (!checkCurrentGameInvariants(world, field)) {
         determineNewPlays(world, field);
-        return false;
+        /// TODO: there should be a call to PlayDecider here, but this is not implemented yet.
     }
 }
 
-const std::vector<Play *> PlayChecker::getValidPlays() const { return validPlays; }
-
 }  // namespace rtt::ai::analysis
-// namespace rtt::ai::analysis
